refactor(web): inline ADMIN_VISIBLE_QUEUES into existing constants.ts
Task 1 initially created a new /constants/admin-queues.ts file, but
/constants.ts already exists (17KB, 20+ importers). Having both a
flat file and a sibling directory under the same alias is a resolution
ambiguity — inline the tiny (25-line) payload into constants.ts to
match the existing convention.

Also fix a JSDoc typo: QueueName enum member is Notifications (plural),
not Notification.
