+++ web/src/lib/components/QueuePanel.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import QueueCard from '$lib/components/QueueCard.svelte';
   import QueueStorageMigrationDescription from '$lib/components/QueueStorageMigrationDescription.svelte';
-  import { ADMIN_VISIBLE_QUEUES, type AdminVisibleQueue } from '$lib/constants/admin-queues';
+  import { ADMIN_VISIBLE_QUEUES, type AdminVisibleQueue } from '$lib/constants';
   import { featureFlagsManager } from '$lib/managers/feature-flags-manager.svelte';
   import { queueManager } from '$lib/managers/queue-manager.svelte';
   import { asQueueItem } from '$lib/services/queue.service';
